Add protected route
