+++ src/components/LoginForm/LoginForm.jsx
@@ -1,10 +1,12 @@
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import React, { useState } from 'react';
 import Input from '../Input/Input';
 import { getToken } from '../../redux/authentication/authenticationSlice';
+import FormErrors from '../ui/FormErrors';
 
 const LoginForm = () => {
   const dispatch = useDispatch();
+  const { error } = useSelector((state) => state.authentication);
   const [username, setUsername] = useState({ target: { value: '' } });
 
   const handleSubmit = (e) => {
@@ -13,10 +15,15 @@ const LoginForm = () => {
   };
 
   return (
-    <form onSubmit={handleSubmit}>
-      <Input name="username" type="text" onInput={setUsername} />
-      <input type="submit" value="login" />
-    </form>
+    <section id="form">
+      <div className="form__container">
+        <FormErrors formErrors={{ error }} />
+        <form onSubmit={handleSubmit}>
+          <Input name="username" type="text" onInput={setUsername} />
+          <input type="submit" value="login" />
+        </form>
+      </div>
+    </section>
   );
 };
