+++ src/components/ui/FormErrors.jsx
@@ -1,12 +1,12 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 
-const FormErrors = ({ formErrors, isValid }) => (
-  <div className={isValid || isValid == null ? '' : 'form__errors'}>
+const FormErrors = ({ formErrors }) => (
+  <div className="form__errors">
     {
       Object.keys(formErrors).map((field) => (
         formErrors[field].length > 0 && (
-        <p key={`${field}`}>
+        <p className="form__error" key={`${field}`}>
           {field}
           :
           {' '}
@@ -20,11 +20,6 @@ const FormErrors = ({ formErrors, isValid }) => (
 
 FormErrors.propTypes = {
   formErrors: PropTypes.objectOf(PropTypes.string).isRequired,
-  isValid: PropTypes.bool,
-};
-
-FormErrors.defaultProps = {
-  isValid: null,
 };
 
 export default FormErrors;
